back-end/requirements.txt
fastapi[standard]
openai
pdfplumber
python-dotenv
langchain
langchain-community